Struct graph_node::store_builder::StoreBuilder

pub struct StoreBuilder {
    logger: Logger,
    subgraph_store: Arc<SubgraphStore>,
    pools: HashMap<ShardName, ConnectionPool>,
    subscription_manager: Arc<SubscriptionManager>,
    chain_head_update_listener: Arc<PostgresChainHeadUpdateListener>,
    chains: HashMap<String, ShardName>,
}

Fields
logger: Loggersubgraph_store: Arc<SubgraphStore>pools: HashMap<ShardName, ConnectionPool>subscription_manager: Arc<SubscriptionManager>chain_head_update_listener: Arc<PostgresChainHeadUpdateListener>chains: HashMap<String, ShardName>

Map network names to the shards where they are/should be stored

Implementations

impl StoreBuilder

pub async fn new(
logger: &Logger,
node: &NodeId,
config: &Config,
registry: Arc<dyn MetricsRegistry>
) -> Self

Set up all stores, and run migrations. This does a complete store setup whereas other methods here only get connections for an already initialized store

pub fn make_subgraph_store_and_pools(
logger: &Logger,
node: &NodeId,
config: &Config,
registry: Arc<dyn MetricsRegistry>
) -> (Arc<SubgraphStore>, HashMap<ShardName, ConnectionPool>)

Make a ShardedStore across all configured shards, and also return the main connection pools for each shard, but not any pools for replicas

pub fn make_subgraph_store(
logger: &Logger,
node: &NodeId,
config: &Config,
registry: Arc<dyn MetricsRegistry>
) -> Arc<SubgraphStore>

pub fn make_store(
logger: &Logger,
pools: HashMap<ShardName, ConnectionPool>,
subgraph_store: Arc<SubgraphStore>,
chains: HashMap<String, ShardName>,
networks: Vec<(String, Vec<EthereumNetworkIdentifier>)>
) -> Arc<DieselStore>

pub fn main_pool(
logger: &Logger,
node: &NodeId,
name: &str,
shard: &Shard,
registry: Arc<dyn MetricsRegistry>
) -> ConnectionPool

Create a connection pool for the main database of hte primary shard without connecting to all the other configured databases

fn replica_pools(
logger: &Logger,
node: &NodeId,
name: &str,
shard: &Shard,
registry: Arc<dyn MetricsRegistry>
) -> (Vec<ConnectionPool>, Vec<usize>)

Create connection pools for each of the replicas

pub fn network_store(
self,
networks: Vec<(String, Vec<EthereumNetworkIdentifier>)>
) -> Arc<DieselStore>

Return a store that combines both a Store for subgraph data and a BlockStore for all chain related data

pub fn subscription_manager(&self) -> Arc<SubscriptionManager>

pub fn chain_head_update_listener(&self) -> Arc<PostgresChainHeadUpdateListener>

pub fn primary_pool(&self) -> ConnectionPool